One-click Installation
Installer is now a one-click installation; no need to have Plays 3.0.0 installed.

The process is now in a nutshell: Download PlaysSetup.exe from web.archive.org. Extract the .exe, modify nupkg package, compile and install.

The release will be packaged in a way where it will only be an executable. It will self extract to '%localappdata%\RePlays' and do all the work there.

This process not only dumbs down the installation, but also speeds up installation. Woo hoo!