Added Search Host
The Search Host runs when searching for something in Windows. It's not a game and also blocks the game detection for other games when trying to get the Window Size.
RePlays will not record the (actual) game if you start the game within 20 seconds after searching for something in Windows.